I still dont know D::::::

+++ include/Template/pure-pursuit/custom-math.h
@@ -13,6 +13,6 @@ bool is_past_segment(Point robot_pos, Point line_point_1, Point line_point_2);
 
 int sgn(float x);
 
-float findLookaheadCurvature(Point robot_pos, float heading, Point lookahead);
+double findLookaheadCurvature(Point robot_pos, float heading, Point lookahead);
 
 float slew(float target, float current, float max_change, int restrict_direction);
