+++ src/Template/pure-pursuit/custom-math.cpp
@@ -95,15 +95,15 @@ int sgn(float x){
  * @param lookahead the lookahead point
  * @return float curvature
  */
-float findLookaheadCurvature(Point robot_pos, float heading, Point lookahead) {
+double findLookaheadCurvature(Point robot_pos, float heading, Point lookahead) {
     // calculate whether the robot is on the left or right side of the circle
-    float side = sgn(sin(to_rad(heading)) * (lookahead.x - robot_pos.x) - cos(to_rad(heading)) * (lookahead.y - robot_pos.y));
+    int side = sgn(sin(to_rad(heading)) * (lookahead.x - robot_pos.x) - cos(to_rad(heading)) * (lookahead.y - robot_pos.y));
     
     // calculate center point and radius
-    float a = -tan(to_rad(heading));
-    float c = tan(to_rad(heading)) * robot_pos.x - robot_pos.y;
-    float x = fabs(a * lookahead.x + lookahead.y + c) / sqrt((a * a) + 1);
-    float d_squared = pow(lookahead.x - robot_pos.x, 2) + pow(lookahead.y - robot_pos.y, 2);
+    double a = -tan(to_rad(heading));
+    double c = tan(to_rad(heading)) * robot_pos.x - robot_pos.y;
+    double x = fabs(a * lookahead.x + lookahead.y + c) / sqrt((a * a) + 1);
+    double d_squared = pow(lookahead.x - robot_pos.x, 2) + pow(lookahead.y - robot_pos.y, 2);
 
     // return curvature
     return side * ((2 * x) / (d_squared));
